requirements.txt
assemblyai
requests
python-dotenv
flask
